feat: picker actions, command module split, cover art fix, dir spider
Picker confirm now has 3 actions:
- Enter: append to queue (don't play)
- Ctrl+Enter: append and play
- Ctrl+R: replace entire queue

Split main.rs subcommands into commands/ directory (12 modules).
Fix cover art black bar by forcing even pixel height.
koan play with directories recursively walks for audio files and
reads metadata for proper album grouping.
