+++ crates/koan-cli/src/commands/cache.rs
@@ -1,0 +1,88 @@
+use koan_core::config;
+use owo_colors::OwoColorize;
+
+use super::{format_bytes, open_db};
+
+pub fn cmd_cache_status() {
+    let cfg = config::Config::load().unwrap_or_default();
+    let cache_dir = cfg.cache_dir();
+
+    println!("{} {}", "path:".cyan(), cache_dir.display());
+
+    if !cache_dir.exists() {
+        println!(
+            "{} {}",
+            "size:".cyan(),
+            "empty (no cache directory)".dimmed()
+        );
+        return;
+    }
+
+    let mut total_bytes: u64 = 0;
+    let mut file_count: u64 = 0;
+    for entry in walkdir::WalkDir::new(&cache_dir)
+        .into_iter()
+        .filter_map(|e| e.ok())
+        .filter(|e| e.file_type().is_file())
+    {
+        if let Ok(meta) = entry.metadata() {
+            total_bytes += meta.len();
+            file_count += 1;
+        }
+    }
+
+    let size = format_bytes(total_bytes);
+    println!(
+        "{} {} {}",
+        "size:".cyan(),
+        size.bold(),
+        format!("({} files)", file_count).dimmed(),
+    );
+}
+
+pub fn cmd_cache_clear() {
+    let cfg = config::Config::load().unwrap_or_default();
+    let cache_dir = cfg.cache_dir();
+
+    if !cache_dir.exists() {
+        println!("{}", "cache already empty".dimmed());
+        return;
+    }
+
+    // Count what we're about to nuke.
+    let mut total_bytes: u64 = 0;
+    let mut file_count: u64 = 0;
+    for entry in walkdir::WalkDir::new(&cache_dir)
+        .into_iter()
+        .filter_map(|e| e.ok())
+        .filter(|e| e.file_type().is_file())
+    {
+        if let Ok(meta) = entry.metadata() {
+            total_bytes += meta.len();
+            file_count += 1;
+        }
+    }
+
+    if file_count == 0 {
+        println!("{}", "cache already empty".dimmed());
+        return;
+    }
+
+    if let Err(e) = std::fs::remove_dir_all(&cache_dir) {
+        eprintln!("{} {}", "error:".red().bold(), e);
+        std::process::exit(1);
+    }
+
+    // Clear cached_path in DB so tracks get re-downloaded next time.
+    let db = open_db();
+    let _ = db
+        .conn
+        .execute("UPDATE tracks SET cached_path = NULL", rusqlite::params![]);
+
+    println!(
+        "{} {} {}",
+        "cache cleared".green().bold(),
+        format_bytes(total_bytes),
+        format!("({} files removed)", file_count).dimmed(),
+    );
+}
