+++ crates/koan-cli/src/commands/remote.rs
@@ -1,0 +1,117 @@
+use koan_core::config;
+use owo_colors::OwoColorize;
+
+use super::{get_remote_password, open_db};
+
+pub fn cmd_remote_login(url: &str, username: &str) {
+    let password = rpassword::prompt_password("password: ").unwrap_or_else(|e| {
+        eprintln!("{} {}", "error:".red().bold(), e);
+        std::process::exit(1);
+    });
+
+    let client = koan_core::remote::client::SubsonicClient::new(url, username, &password);
+    match client.ping() {
+        Ok(()) => println!("{} {}", "connected".green(), url),
+        Err(e) => {
+            eprintln!("{} {}", "connection failed:".red().bold(), e);
+            std::process::exit(1);
+        }
+    }
+
+    // Load existing local config (preserve folders etc), update remote fields.
+    let local_path = config::config_local_file_path();
+    let mut local_cfg = if local_path.exists() {
+        config::Config::load_from(&local_path).unwrap_or_default()
+    } else {
+        config::Config::default()
+    };
+    local_cfg.remote.enabled = true;
+    local_cfg.remote.url = url.to_string();
+    local_cfg.remote.username = username.to_string();
+    local_cfg.remote.password = password;
+    if let Err(e) = local_cfg.save_local() {
+        eprintln!("{} {}", "config error:".red().bold(), e);
+        std::process::exit(1);
+    }
+    println!("{}", "credentials saved to config.local.toml".green());
+}
+
+pub fn cmd_remote_sync() {
+    let cfg = config::Config::load().unwrap_or_default();
+    if !cfg.remote.enabled || cfg.remote.url.is_empty() {
+        eprintln!(
+            "{} no remote server configured — run {} first",
+            "error:".red().bold(),
+            "koan remote login".bold()
+        );
+        std::process::exit(1);
+    }
+
+    let password = get_remote_password(&cfg);
+
+    let client = koan_core::remote::client::SubsonicClient::new(
+        &cfg.remote.url,
+        &cfg.remote.username,
+        &password,
+    );
+
+    let db = open_db();
+    let start = std::time::Instant::now();
+
+    match koan_core::remote::sync::sync_library(&db, &client) {
+        Ok(result) => {
+            let elapsed = start.elapsed();
+            println!(
+                "{} {} {} artists, {} albums, {} tracks",
+                "sync complete".green().bold(),
+                format!("({:.1}s)", elapsed.as_secs_f64()).dimmed(),
+                result.artists_synced.to_string().bold(),
+                result.albums_synced.to_string().bold(),
+                result.tracks_synced.to_string().bold(),
+            );
+        }
+        Err(e) => {
+            eprintln!("{} {}", "sync failed:".red().bold(), e);
+            std::process::exit(1);
+        }
+    }
+}
+
+pub fn cmd_remote_status() {
+    let cfg = config::Config::load().unwrap_or_default();
+    if !cfg.remote.enabled || cfg.remote.url.is_empty() {
+        println!("no remote server configured");
+        return;
+    }
+
+    println!("{} {}", "server:".cyan(), cfg.remote.url);
+    println!("{} {}", "username:".cyan(), cfg.remote.username);
+
+    let has_password = !cfg.remote.password.is_empty();
+    println!(
+        "{} {}",
+        "password:".cyan(),
+        if has_password {
+            "configured".green().to_string()
+        } else {
+            "not set".red().to_string()
+        }
+    );
+
+    if has_password {
+        let client = koan_core::remote::client::SubsonicClient::new(
+            &cfg.remote.url,
+            &cfg.remote.username,
+            &cfg.remote.password,
+        );
+        match client.ping() {
+            Ok(()) => println!("{} {}", "status:".cyan(), "connected".green()),
+            Err(e) => println!(
+                "{} {} {}",
+                "status:".cyan(),
+                "error".red(),
+                format!("\u{2014} {}", e).dimmed()
+            ),
+        }
+    }
+}
